Clean up the docker dev env. All Braven docker containers are in the same network now.
This is so that from within the container, it can call out to another
container not in this docker-compose.yml file and the service name will
resolve. E.g. http://canvasweb would resolve. And vice version, from
canvasweb, it can call to http://cssjsweb and it will resolve when on
the same network.

TESTING:
- Rebuilt the container, then connected to the ssoweb one in this same
  network and ran: curl http://cssjsweb:3004/bz_custom.js which worked.