Make inline rubric loading fast.
For each inline rubric row, the old code would make an AJAX call for the
entire rubric and then pull that row out. For an assignment with a lot of
inline rubrics, this could end up making like 20-30 AJAX requests for the same
full rubric (each taking over 100ms, with some up to 500+ms).
For example, I ran a load test simulating 5 Fellows loading the Hustle To Career
project 4 times in a row with different groups of 4. It resulted in 768 calls to the
rubric controller for a total wait time of over a minute just on rubrics.
This new code only makes the AJAX request for the full rubric once and then uses
that to replace each inline rubric row.

TESTING:
- Loaded http://canvasweb:3000/courses/58/assignments/1314 and looked at the console
  log to make sure the AJAX request only happened once. Also verified that the inline
  rubrics actually worked.
- Loaded http://canvasweb:3000/courses/10/pages/ajax-load-rubric-on-page to verify that clickable
  rubrics still work. Also, renamed the clickable rubric magic field to similuate a first time
  load where there is no backing magic field set yet.
- Alternated between the two places that actually do the rubric replace and put in a sleep(15)
  to force the other one to run first, ensuring that no matter which side wins the race condition
  both work.

POST-RELEASE
- Re-run the load test (4 groups of 5 Fellows loading Hustle to Career Porject) and compare
  the Honeycomb results to this:
  https://ui.honeycomb.io/braven/datasets/stagingportal.bebraven.org/result/A41LshAE2Sw