Style the button where students click to see annotated file feedback.
When a TA annotates a file in an assignment submission, the Fellow
sees the file with this tiny little link that says "View Feedback"
which is hard to find. This just re-style's the link to be a Call To Action
button.

See: https://app.asana.com/0/1174274412967132/1199154572811210

To release this it must be uploaded to the Braven theme on braven.instructure.com