Customize the page where you launch an LTI assignment into a new window.
We're switching our Modules and Projects to have the "Open in new tab"
setting on. With this on, when you click on a Module or Project
you see a page with a button asking you to click the button to open
the "tool" in a new window. This customizes the language there to not
use "tool" since Fellows know these as assignments. Also, add a little
whitespace to make it look better.